move cron jobs to folder and add schedule for upcoming featured show check
