+++ vercel.json
@@ -4,8 +4,12 @@
   },
   "crons": [
     {
-      "path": "/api/show-submission-email",
+      "path": "/api/cron/show-submission-email",
       "schedule": "0 9 * * *"
+    },
+    {
+      "path": "/api/cron/upcoming-featured-show-check",
+      "schedule": "27 9 * * *"
     }
   ]
 }
